Split `AddPresetPopover` to separate file
